still adressing bug reported by Ihab

+++ Arduino/fly_pi_development/fly_pi_development.ino
@@ -141,7 +141,7 @@ void loop() {
   //check temp sensor
   if (term1 == "TEM") 
     {  temperature = checkTemp(tempSensorPin);
-        newTemp=term2;
+        newTemp=term2.toInt();
        Serial.println(temperature);
        if (peltOn == 1) {
           HoldTemp(newTemp, tempSensorPin, peltierCoolPin1, peltierHeatPin1);
